pyproject.toml
[project]
name = "embod"
version = "0.1.0"
description = "AI-native CADQuery embodiment tooling"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Embod" }]
dependencies = [
  "pydantic>=2.11.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "rich>=14.0.0,<15.0.0",
  "typer>=0.16.0,<1.0.0",
  "cadquery>=2.7.0,<3.0.0",
]

[project.optional-dependencies]
viz = [
  "numpy>=2.2.0,<3.0.0",
  "pillow>=11.1.0,<12.0.0",
  "pyvista>=0.46.0,<1.0.0",
]
sim = [
  "numpy>=2.2.0,<3.0.0",
  "pillow>=11.1.0,<12.0.0",
]

[dependency-groups]
dev = [
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.4.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "ruff>=0.11.0,<1.0.0",
  "types-PyYAML>=6.0.12.20241230,<7.0.0",
]

[project.scripts]
embod = "embod.cli.main:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embod"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "C4",
  "UP",
  "ARG",
  "SIM",
  "ANN",
  "RUF",
]
ignore = [
  "ANN101",
  "ANN102",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001"]

[tool.mypy]
python_version = "3.11"
mypy_path = ["stubs", "src"]
strict = true
warn_return_any = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_any_unimported = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.uv]
default-groups = ["dev"]
